requirements.txt
ipython
Django>=3.1,<4.0
psycopg2-binary>=2.8
celery>=5.0,<6.0
redis==3.5.3
boto3>=1.16
gunicorn>=20.0
django-redis>=3.3
django-celery-beat==2.1.0
django-admin-rangefilter>=0.6.3,<0.7
django-ckeditor>=6.0
django-extensions==3.0.9
django-fsm>=2.7.1,<2.8
django-ses>=1.0
django-taggit==1.3
django-allauth>=0.43
django-countries>=6.1,<6.2
django-storages>=1.10,<1.20
django-filter==2.4.0
pillow>=8.0
itsdangerous==2.0.0a1
isort>=5.6.4
coverage==5.3
stripe==2.55.0
django-debug-toolbar>=3.1,<4.0
django-jazzmin==2.4.0
Werkzeug>=1.0.1
django-import-export==2.4.0
factory-boy>=3.1.0,<3.2
django-oauth-toolkit>=1.3.3,<1.4
django-widget-tweaks>=1.4.8,<1.5
django-cors-headers>=3.5.0,<3.6
django-simpleui
django-mathfilters>=1.0.0,<1.1
django-mptt>=0.11.0